refactor(accessor): move accessor to /utils
