+++ src/utils/domain.ts
@@ -1,6 +1,6 @@
 import { extent, sum } from 'd3-array';
 import { nest } from 'd3-collection';
-import { Accessor, AccessorFn } from '../lib/utils/accessor';
+import { Accessor, AccessorFn } from './accessor';
 import { ScaleType } from './scales/scales';
 
 export type Domain = any[];
